Adicionando validação no componente Livewire UserList para previnir edição ou deleção de usuários não permitidos

+++ app/Http/Livewire/UserList.php
@@ -5,6 +5,7 @@ namespace App\Http\Livewire;
 use Livewire\Component;
 use Livewire\WithPagination;
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 
 class UserList extends Component
 {
@@ -27,11 +28,23 @@ class UserList extends Component
 
     public function buttonUpdate($id)
     {
+        if(!Auth::check() ||  Auth()->user()->id != $id)
+        {
+            session()->flash('message', 'Não é possível editar esse usuário!');
+            return;
+        }
+
         return $this->redirect('/update/'.$id);
     }
 
     public function buttonDelete($id)
     {
+        if(!Auth::check() ||  Auth()->user()->id != $id)
+        {
+            session()->flash('message', 'Não é possível deletar esse usuário!');
+            return;
+        }
+
         if($id){
             User::where('id',$id)->delete();
             session()->flash('message', 'Usuário deletado com sucesso!');
